Fix some issues with file selection
There was a bug after pressing the clear selected button. Re-selecting a file
would add it to the old list of selected files that were never cleared.
There was also a bug in the behavior of the green "files selected" message.
It should only appear when files have actually been added to
the selection queue, and not just after the "Select" button is pressed.